WIP: backup before switching to adoptPage
