+++ src/components/common/adoptCard/AdoptCard.css
