+++ src/components/common/adoptCard/AdoptCard.jsx
