+++ src/components/common/communityCard/CommunityCard.css
