--- src/pages/AdoptSelect.css
@@ -1,98 +1,0 @@
-/* AdoptSelect.css */
-
-.adopt-select-wrapper {
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-}
-
-.adopt-select-body {
-  display: flex;
-  flex: 1;
-  align-items: stretch;
-}
-
-/* ✅ 사이드바: Footer까지 이어지고, 카드 첫줄과 정렬 */
-.left-sidebar {
-  width: 228px;
-  background-color: #d9d9d9;
-  min-height: 100vh;         /* Footer까지 채우기 */
-  margin-top: 88px;          /* Navbar 높이 고려하여 카드와 맞춤 */
-  box-sizing: border-box;
-}
-
-/* ✅ 메인 콘텐츠 */
-.adopt-select-container {
-  flex: 1;
-  padding: 2rem 0;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-/* 타이틀 */
-.adopt-title {
-  font-size: 2rem;
-  font-weight: bold;
-  text-align: center;
-  margin: 0 auto 0.5rem auto;
-}
-
-.adopt-title-underline {
-  height: 1px;
-  background-color: #999;
-  width: 160px;
-  margin: 0 auto 2rem auto;
-}
-
-/* 카드 영역 */
-.adopt-card-box {
-  width: 100%;
-  max-width: 1120px;
-  padding: 1.5rem;
-  box-sizing: border-box;
-  border: none;
-  border-radius: 0;
-}
-
-.adopt-card-grid {
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  gap: 1.5rem;
-  justify-items: center;
-}
-
-/* 페이지네이션 */
-.pagination {
-  margin-top: 2rem;
-  display: flex;
-  justify-content: center;
-  gap: 0.5rem;
-}
-
-.page-btn.circle {
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  border: 1px solid #a855f7;
-  color: #6b21a8;
-  background-color: #f3e8ff;
-  font-weight: 500;
-  cursor: pointer;
-  transition: 0.2s ease;
-}
-
-.page-btn.circle:hover:not(:disabled) {
-  background-color: #e9d5ff;
-}
-
-.page-btn.circle.active {
-  background-color: #c084fc;
-  color: white;
-  font-weight: bold;
-}
-
-.page-btn.circle:disabled {
-  opacity: 0.3;
-  cursor: default;
-}
