--- src/pages/AdoptionPage.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-const AdoptionPage = () => {
-  return <div>임시 페이지입니다. 추후 병합 시 대체됩니다.</div>;
-};
-export default AdoptionPage;
