+++ src/routes/adoptPage/AdoptPage.css
@@ -12,110 +12,110 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  margin-top: -40px;
+  margin-top: -20px;
 }
 
 .adopt-title {
-  font-size: 36px;
+  font-size: 2.5rem;
   font-weight: bold;
-  margin-bottom: 6px;
+  text-align: center;
+  margin: 0 auto 0.5rem auto;
 }
 
-.underline {
-  width: 80px;
-  height: 3px;
-  background-color: #000;
-  margin-bottom: 40px;
+.adopt-title-underline {
+  height: 2px;
+  background-color: #999;
+  width: 200px;
+  margin: 0 auto 2rem auto;
 }
 
 .adopt-body-wrapper {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: flex-start;
-  gap: 100px;
-  width: 100%;
+  display: grid;
+  grid-template-columns: auto 1fr auto; /* 썸네일+메인 | 글 | Info */
+  gap: 150px;
   max-width: 1400px;
-  padding-left: 80px;
+  width: 100%;
+  margin-top: 80px;
+  align-items: flex-start;
+  justify-content: center;
 }
 
-.left-group {
+/* 썸네일 + 메인이미지 */
+.image-group {
   display: flex;
-  gap: 32px;
-  flex: 1;
-  margin-left: -215px;
+  flex-direction: row;
+  align-items: flex-start;
+  gap: 20px;
+  grid-column: 1;
 }
 
 .thumbnail-column {
   display: flex;
   flex-direction: column;
-  gap: 16px;
-}
-
-.thumbnail-img {
-  width: 60px;
-  height: 60px;
-  border-radius: 10px;
-  object-fit: cover;
-  background-color: #ccc;
-}
-
-.image-info-group {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
   gap: 20px;
 }
 
+.thumbnail-img {
+  width: 100px;
+  height: 100px;
+  border-radius: 10px;
+  object-fit: cover;
+  background-color: #ccc;
+  transition: transform 0.2s ease;
+}
+
+.thumbnail-img:hover {
+  transform: scale(1.05);
+}
+
 .main-img {
-  width: 400px;
-  height: 400px;
+  width: 450px;
+  height: 450px;
   object-fit: cover;
   border-radius: 20px;
   background-color: #666;
-  margin-bottom: 12px;
 }
 
+/* 가운데 글 영역 */
 .text-column {
   display: flex;
   flex-direction: column;
-  justify-content: space-between; /* 위-아래로 나눔 */
   align-items: flex-start;
-  height: 380px;  /* 더 키움 → Chat이 아래로 내려가게 */
-  text-align: left;
-  font-size: 20px;
-  margin-left: 180px;
+  gap: 20px;
+  font-size: 24px;
+  width: 300px;
+  grid-column: 2;
 }
 
 .row-inline {
   display: flex;
   align-items: center;
-  gap: 12px;
+  gap: 20px;
 }
 
 .heart-button {
   background: none;
   border: 1px solid #999;
   border-radius: 10px;
-  padding: 6px 10px;
-  margin-left: 10px;  /* ✅ 간격 추가 */
-  transform: scale(1.4); /* ✅ 버튼 크기 키우기 */
-  font-size: 20px;
+  padding: 8px 12px;
+  transform: scale(1.5);
+  font-size: 22px;
   cursor: pointer;
 }
 
 .chat-row {
   display: flex;
   align-items: center;
-  gap: 10px;
-  margin-top: auto;
+  gap: 14px;
+  margin-top: 16px;
 }
 
-.chat-gradient-circle {
-  width: 48px;
-  height: 48px;
+.chat-profile-img {
+  width: 60px;
+  height: 60px;
   border-radius: 50%;
-  background: linear-gradient(to right, #ff0080, #7928ca);
+  object-fit: cover;
+  background-color: #eee;
 }
 
 .chat-icon {
@@ -125,34 +125,35 @@
 .chat-text {
   background: none;
   border: none;
-  font-size: 24px;
+  font-size: 26px;
   font-weight: 600;
   cursor: pointer;
 }
 
+/* 오른쪽 설명 영역 */
 .description-column {
   display: flex;
   flex-direction: column;
-  align-items: flex-end; /* ✅ 모든 내부 요소 오른쪽 정렬 */
-  width: 100%;
-  max-width: 420px;
-  margin-right: 40px; /* ✅ 사진/글씨와의 간격 조정 */
+  align-items: flex-start;
+  justify-content: flex-start; 
+  width: 400px;
+  margin-top: -20px;          
 }
 
 .description-column h3 {
   color: #2c2cce;
-  font-size: 20px;
+  font-size: 24px;
   font-weight: bold;
-  align-self: flex-start;
+  margin-bottom: 10px;
 }
 
 .description-column textarea {
-  width: 100%; /* ✅ 부모 기준 100%로 유지 */
-  height: 400px;
-  font-size: 16px;
+  width: 100%;
+  height: 500px;
+  font-size: 18px;
   border: 2px solid #814ffb;
   border-radius: 6px;
-  padding: 12px;
+  padding: 14px;
   font-family: inherit;
   resize: none;
   box-sizing: border-box;
@@ -163,12 +164,12 @@
   background-color: #814ffb;
   color: white;
   border: none;
-  padding: 12px 28px;
+  padding: 14px 30px;
   font-weight: bold;
-  font-size: 16px;
+  font-size: 18px;
   border-radius: 8px;
-  width: 120px;
-  margin-top: 12px;
+  width: 130px;
+  margin-top: 16px;
   cursor: pointer;
 }
