+++ src/routes/adoptPage/AdoptSelect.css
@@ -1,5 +1,3 @@
-/* AdoptSelect.css */
-
 .adopt-select-wrapper {
   display: flex;
   flex-direction: column;
@@ -58,7 +56,7 @@
 .adopt-card-grid {
   display: grid;
   grid-template-columns: repeat(4, 1fr);
-  gap: 1.5rem;
+  gap: 5.5rem;
   justify-items: center;
 }
