+++ src/routes/communityPage/CommunityGuideline.jsx
@@ -1,16 +1,13 @@
 import React from 'react';
-import '../pages/CommunityGuideline.css';
-import Navbar from "../components/common/navbar/Navbar";
-import Footer from "../components/common/footer/Footer";
-import { useLanguage } from "../context/language/LanguageContext";
-import dogPic from '../assets/dog_pic.png';
+import "./CommunityGuideline.css"; 
+import { useLanguage } from "../../context/language/LanguageContext";
+import dogPic from "../../assets/dog_pic.png";  
 
 const CommunityGuideline = () => {
   const { language } = useLanguage();
 
   return (
     <div className="guideline-wrapper">
-      <Navbar />
       <div className="guideline-container">
         <div className="main-layout-wrapper">
           <div className="guideline-title-wrapper">
@@ -84,7 +81,6 @@ const CommunityGuideline = () => {
           </div>
         </div>
       </div>
-      <Footer />
     </div>
   );
 };
